+++ lib/index.d.ts
@@ -25,5 +25,14 @@ export declare const Commands: (
       sap_UoMs_total: number;
       repzo_products_total: number;
     }
+  | {
+      repzo_PL_total: number;
+      sap_PL_items_total: number;
+      repzo_PL_items_total: number;
+      repzo_PL_items: {
+        removed: number;
+        failed: number;
+      };
+    }
 >;
 export declare const CommandsList: import("./types.js").Command[];